requirements.txt
torch>=2.1.0
numpy>=1.24
nibabel>=5.2
scipy>=1.10
SimpleITK>=2.3
monai>=1.3
fastapi>=0.115
uvicorn>=0.30
onnxruntime>=1.18
reportlab>=4.2
Pillow>=10.0
python-multipart>=0.0.9
requests>=2.32
